Separated authentication function into a separate model and improved authentication error catch
